Argparse, modular code, and config handling
This commit addresses the following:
fixes #9 - Fully implements argparse, no more legacy manual parsing
fixes #10 - Runs initial config file creation when 'ncc' command is initially run
fixes #4 - Config generation also runs when 'ls' is used
fixes #6 - Implements better error handling for invalid URLs and provided helpful message

This commit also adds the following:
- More modular code, no longer a massive monolithic file which is difficult to maintain
- Adds .env file for better pylint behavior in VSCode editor
- Adds launch.json file for better debug behavior in VSCode
