+++ ncc_cli/__init__.py
@@ -1,2 +1,5 @@
 name = "ncc_cli"
 from .ncc_cli import *
+from .connection import *
+from .transfer import *
+from .commands import *
