+++ ncc_cli/utils.py
@@ -1,0 +1,11 @@
+# This file defines utility functions
+
+# Placeholder debugPrint function, replace with useful code later
+def debugPrint(message):
+    print(message)
+
+#Define some pretty colours to be used later
+class bcolours: 
+    BLUE = '\033[34m'
+    DEFAULT = '\033[39m'
+    YELLOW = '\033[93m'
